feat: Add custom transform dialog for user-defined data transformations
- Introduced a new `CustomTransformDialog` class to allow users to create custom data transformations by writing mathematical formulas.
- Implemented formula input with live syntax validation and error feedback.
- Added sections for available variables and functions to assist users in constructing formulas.
- Included a preview feature to display the transformation results for the first 8 data points.
- Integrated the dialog into the ImGui UI, allowing it to be opened from the command bar.
- Created an expression evaluation system to parse and compute the results of user-defined formulas.
- Added necessary data structures and methods for managing series data and context during evaluation.

+++ include/spectra/eigen.hpp
@@ -305,7 +305,7 @@ inline vec3 to_vec3(const Eigen::Vector3f& v)
 // Convert spectra::vec3 to Eigen::Vector3f
 inline Eigen::Vector3f to_eigen(const vec3& v)
 {
-    return {v.x, v.y, v.z};
+    return {static_cast<float>(v.x), static_cast<float>(v.y), static_cast<float>(v.z)};
 }
 
 // Convert Eigen::Matrix4f (column-major) to spectra::mat4
